requirements.txt
django
django-rest-knox
djangorestframework
